feat: enforced good board geometry using FreeCAD
Parametric modeling is a wonderful thing and I'm unclear on why
KiCAD doesn't have it. Oh well, it is what is is. Using the
dimensions I obtained from the wonderful experience of performing
my parametric modeling adjustments, I finally had all of the
information I needed to get pricing for the PCBs as they are
currently designed. The results were not good. Currently,
[Aisler](https://aisler.community/t/pcb-design-rules/41) rates the
design as complex. This fails to meet the mission goal of Keybert
in a couple of ways.

Firstly, a complex PCB is an expensive PCB and an expensive PCB is
not beginner friendly. Secondly, a complex PCB is a harder to
understand and evaluate PCB and a harder to understand and evaluate
PCB is a PCB that doesn't help a biginner wrap their heads around
the sorts of decisions are present in the general world of custom
mechanical keyboards. Also, as an open source project, I feel a
certain degree of responsibility to make sure that this is as easy
to remix and remaster as possible. I look at the ecosystem left by
wootapoot and am so glad that the Let's Split was open sourced.
That board was a wonderful and excellent board, but the original
creator has seemingly vanished. I want when I stop being available
to support Keybert for Keybert to be able to venture out into the
world supported by many, just like the Let's Split lives on and
even has inspired things like the Lundquist and Let's Tango

chore: tidy up repository

Look at this place! Look at this mess! This isn't easy to understand or
explore! What are we even doing here!?

fix: moved file into correct directory

I accidentally broke all of the KiCAD files with the last commit. Now
things should be better and functional