fix: No longer incurring complexity fee
After a lot of work, the PCB itself should now be viable for ordering.
At time of completion of that work, the PCB is $30 to be produced in the
United states in a batch of 3 or $22 to be produced in China in a batch
of 5. Arguably, that batch of five is more in line with the vision for
Keybert, however, I would forsee myself preferring the American PCB
manufacturer for my own boards given overall sociopolitical
circumstances in the world. I have decided that Keybert will exist
purely in the cheapest possible way with no guidance on how ethical or
unethical it is to engage with various entities. There likely will need
to be documentation somewhere stating outright that the creator of
Keybert in no way condones things like the Uigher genocide, the invasion
of Ukraine, or any other form of genocide

fix: added library file for TRRS jack

In order to properly work with the Keebio TRRS symbol, I needed to
create a footprint that matched it in terms of pads. Or wait no.
Otherway. I wanted to use the Keebio footprint, so I created a new
symbol to match in terms of pads